+++ packages/data-model/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @mlightcad/data-model
 
+## 1.2.12
+
+### Patch Changes
+
+- feat: add one flag to delay creating one rendered entity in function AcDbEntity.draw
+
+## 1.2.11
+
+### Patch Changes
+
+- feat: attach entity info (such as objectId, layerName and etc.) in rendering cache entities
+
 ## 1.2.8
 
 ### Patch Changes