+++ packages/libredwg-converter/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @mlightcad/libredwg-converter
 
+## 3.0.12
+
+### Patch Changes
+
+- feat: add one flag to delay creating one rendered entity in function AcDbEntity.draw
+- Updated dependencies
+  - @mlightcad/data-model@1.2.12
+
+## 3.0.11
+
+### Patch Changes
+
+- feat: attach entity info (such as objectId, layerName and etc.) in rendering cache entities
+- Updated dependencies
+  - @mlightcad/data-model@1.2.11
+
 ## 3.0.9
 
 ### Patch Changes